Profanify V4 (#48)
# 👋 Introduction
This PR is the next version of Profanify and actually how I wanted it to
work from the beginning. Instead of being a Pest Expectation that you
use in a test file, it takes inspiration from the [Type
Coverage](https://github.com/pestphp/pest-plugin-type-coverage) plugin
and allows you to run it via CLI.

# 🚗 Usage
How you use Profanify has completely changed which is why this PR will
trigger a new major release of the package. Any applications using it
will need to remove the Expectation and instead add it to their CI/CD
pipelines and/or `composer.json` file. To run it, you now need to run:

```
./vendor/bin/pest --profanity
```

I'd highly recommend setting this up as a script in `composer.json`
though:

```json
"scripts": {
    "profanity": "pest --profanity"
}
```

In doing so, you can then run it like this:

```
composer profanity
```
# 🔭 Results
The output is wonderfully simple and clearly shows what profanity (if
any) has been found in each file, along with the line number, so you can
easily track down and remove them. There's also a total at the bottom
which also shows the total number of files that have profanity:

<img width="1100" alt="Screenshot 2025-05-23 at 02 24 50"
src="https://github.com/user-attachments/assets/1ab3c1c3-aa77-43bd-987f-b4a051e52614"
/>


# ⚙️ Additional Options
Profanify V4 comes with various options that you can pass into the
command you run. Let's take a look at them below:

## Specific Language
```
./vendor/bin/pest --profanity --language=en
```
This will only flag profanity that is in the config file of the language
you specify. If you specify a language that doesn't have a config file,
then an error will be shown in the console. Not specifying a language
means that all config files will be checked against.

## Include Words
```
./vendor/bin/pest --profanity --include=elephpant
```
This is a comma separated list of words specific to your application
that you want to flag. This doesn't have to be profanity, this could be
any word that for whatever reason, you don't want to show in your
codebase.

## Exclude Words
```
./vendor/bin/pest --profanity --exclude=elephpant
```
This is a comma separated list of words that although may be present in
your codebase, you may choose to not have them flagged as profanity

## Compact View
```
./vendor/bin/pest --profanity --compact
```
By default, all files will be shown in the output regardless of whether
they contain profanity or not. Depending on how big your application is,
the output could be huge. Using this flag, you can minimise the output
to only show files _with_ profanity.

## 🧾 Output
```
./vendor/bin/pest --profanity --output=test.json
```
If needed, you may send the output to a file by specifying the file name
as the value for this option. The output shows the file path and an
array of profanity found in the file

# 🧪 Testing
The test suite has been updated to cater for each option (and no
options). I've also done some additional testing on some of my
applications to confirm these changes work on a real life application.

# 📓 Additional Notes
- There's no `--min` option. I thought about adding it but I don't think
you'd ever set a minimum profanity %. It doesn't make sense. You either
don't care about profanity in your codebase OR don't want _any_ in your
codebase.
- There's no more list of tolerated words to maintain. The regex has
been updated to better match words which meant the need for it was
redundant.